backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure SDK packages
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-core==1.35.0

# OpenAI and AI services
openai==1.3.7

# Authentication and security
PyJWT[crypto]==2.10.1
cryptography==46.0.2

# HTTP and API
requests==2.32.5
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Date and time handling
python-dateutil==2.8.2

# Type hints
typing-extensions==4.15.0